// Downloads/dpm-ui-web-and-mobile-development/web-apps/consumer-portal/src/types/quoteAndBuy.ts
  export interface VehicleDetails {
    plateNumber?: string;
    registrationYear?: number;
    chassisNo?: string;
    majorColor?: string;
    engineSize?: string;
    transmission?: string;
    vehicleDefinitionType?: string;
    vehicleSequenceNo?: string;
    vehicleCustomID?: string;
    features?: {
      adaptiveCruiseControl: boolean;
      cruiseControl: boolean;
      modification: boolean;
    };
    camera?: {
      rearParkingSensor: boolean;
      frontSensor: boolean;
      frontCamera: boolean;
      rearCamera: boolean;
      degreeCamera: boolean;
    };
    safety?: {
      parking: string;
      antiTheftAlarm: string;
      antiLockBrakingSystem: number | string;
      automaticBrakingSystem?: boolean;
    };
    commercialVehicle?: {
      vehicleAxleWeight: number;
      fireExtinguisher: boolean;
    };
  }

  export interface VehicleDetailsResponseData {
    chassisNumber: string;
    logId: string;
    vehicleColour: string;
    manufactureYear: string;
    vehicleOwnerName: string;
    vehicleRegistrationCity: string;
    vehicleLoad: number | null;
    vehicleMakeText: string;
    vehicleModelText: string | null;
    weight: string;
    vehicleOwnerId: string;
    lkVehBodyType: string;
    lkVehicleClassDesc: string;
    vehicleCapacity: string;
    vehicleMakerCode: string;
    vehicleModelCode: string;
    cylinderCapacity: number | null;
    coOwnerFullName: string | null;
    coOwnerId: string | null;
    plateNo: string | null;
    plateNoText1: string | null;
    plateNoText2: string | null;
    plateNoText3: string | null;
    plateType: string | null;
    regCityLocationCode: string | null;
    vehicleRegistrationExpiryDateH: Date | null;
    transmissionType: string;
    engineCapacity: string;
    referenceNo: string;
    admeid: number;
    description: string;
    vehicleValue: number;
    vehicleMaxValue: number;
    vehicleMinValue: number;
    modelYearID: number;
    make: string;
    makeID: number;
    model: string;
    modelID: number;
    spec: string;
    specID: number;
    engineSizeID: number;
    vehicleCylinders: string;
    noOfCylsID: number;
    hp: string;
    hpID: number;
    vehicleBodyType: string;
    bodyTypeID: number;
    transmissionID: number;
    doors: string;
    doorsID: number;
    seats: string;
    seatsID: number;
    finalDrive: string;
    finalDriveID: number;
    vehicleType: string;
    vehicleTypeID: number;
    fuelType: string;
    fuelID: number;
    engineType: string;
    engineTypeID: number;
    axels: string;
    axelsID: number;
    gears: string;
    gearsID: number;
    cubicCapacity: string;
    cubicCapacityID: number;
    additionFeatures?: VehicleDetails;
    vehicleDefinitionType?: string;
    vehicleCustomID?: string;
  }

  export interface DriverDetails {
    maritalStatusCd?: number;
    childrenBelow16?: number;
    driverRelationship?: string | null;
    educationLevel?: number;
    licenseCountry?: string;
    trafficViolations?: string;
    healthConditions?: string;
    najmCaseDetails?: any[];
    validDrivingLicenses?: any[];
  }

  export interface DriverDetailsResponseData {
    driverID: string;
    driverIDType: string | null;
    driverName: string;
    driverNameArabic: string;
    dateofBirth: string;
    dateofBirthH: string;
    nationality: string;
    gender: string;
    occupation: string;
    workCompanyName: string | null;
    workCityCode: string | null;
    homeCityCode: string | null;
    homeAddress: string;
    licenseType: number | null;
    licenseYear: number | null;
    licenseExpiryDateH: string | null;
    idIssuePlaceCode: string;
    noOfAccidents: number;
    noOfClaims: number;
    unitNo: string | null;
    buildingNumber: string;
    streetName: string;
    district: string;
    city: string;
    additionalNumber: string;
    postalCode: string;
    healthConditions: string | null;
    trafficViolations: string | null;
    najmCaseDetails?: {[key:string]: string}[];
    validDrivingLicenses?: string[];
    mainDriverInd: string;
    ncdFreeYears: string | null;
    ncdReferenceNo: string | null;
    relation?: string | null;
    relationship?: string | null;
    additionalDriverDetails?: DriverDetails
  }

  export interface OwnerDetailsResponseData {
    message: string,
    isValid: boolean,
    referenceNo: number,
    sessionSecretId: string,
    ownerFullNameEnglish: string,
    ownerFullNameArabic: string,
    ownerDobG: string,
    ownerDobH: string,
    gender: string,
    nationality: string,
    nationalityCode: string,
    mobileNumber?: string,
    ownerId?: string,
    nationalId?: string,
    mobile?: string,
  }